chore(actions): do NOT build arm64 on python3.8 and earlier
It seems that arm64 builds are not well-supported on these platforms so
we will skip building the arm64 version on these, for now.

Note that he glob pattern for 3.8 has to be a bit more permissive as the
cibuild pattern aren't fully respected on on github actions for some
reason. The patterns, filters and 'skip' values work correctly works
locally for `cp3{6,7,8}-*-*arm64` bit on github does not match
`cp38-macosx_arm64` - even accounting for the underscore, although the
glob does work for cp36 and cp37... I don't want to be in this position
where I'm debugging `cibuilwheel`.